:recycle: ref: create issue sync spec for github/ghe
